Сделал часть 7 урока
