+++ VKApp/VKApp/Controllers/LoginViewController.swift
@@ -94,7 +94,7 @@ final class LoginViewController: UIViewController {
 
     // MARK: - Private Methods
 
-    @objc private func keyboardWillShown(notification: Notification) {
+    @objc private func keyboardWillShownAction(notification: Notification) {
         guard let info = notification.userInfo as? NSDictionary,
               let keyboard = info.value(forKey: UIResponder.keyboardFrameEndUserInfoKey) as? NSValue
         else { return }
@@ -103,12 +103,12 @@ final class LoginViewController: UIViewController {
         mainScrollView.scrollIndicatorInsets = contectInsets
     }
 
-    @objc private func keyboardWillHide(notification: Notification) {
+    @objc private func keyboardWillHideAction(notification: Notification) {
         mainScrollView.contentInset = UIEdgeInsets.zero
         mainScrollView.scrollIndicatorInsets = UIEdgeInsets.zero
     }
 
-    @objc private func hideKeyboard() {
+    @objc private func hideKeyboardAction() {
         mainScrollView.endEditing(true)
     }
 
@@ -172,21 +172,25 @@ final class LoginViewController: UIViewController {
         return true
     }
 
+    private func addTapToMainScrollView() {
+        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboardAction))
+        mainScrollView.addGestureRecognizer(tapGesture)
+    }
+
     private func keyboardManager() {
         NotificationCenter.default.addObserver(
             self,
-            selector: #selector(keyboardWillShown(notification:)),
+            selector: #selector(keyboardWillShownAction(notification:)),
             name: UIResponder.keyboardWillShowNotification,
             object: nil
         )
         NotificationCenter.default.addObserver(
             self,
-            selector: #selector(keyboardWillHide(notification:)),
+            selector: #selector(keyboardWillHideAction(notification:)),
             name: UIResponder.keyboardWillHideNotification,
             object: nil
         )
-        let tapGesture = UITapGestureRecognizer(target: self, action: #selector(hideKeyboard))
-        mainScrollView.addGestureRecognizer(tapGesture)
+        addTapToMainScrollView()
     }
 
     private func removeKeyboardObserver() {
@@ -201,23 +205,5 @@ extension LoginViewController: UITextFieldDelegate {
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         textField.resignFirstResponder()
         return true
-    }
-}
-
-// MARK: - Alert
-
-extension UIViewController {
-    // MARK: - Constants
-
-    private enum Constants {
-        static let okText = "OK"
-    }
-
-    // MARK: - Public Methods
-
-    func showAlert(title: String, message: String) {
-        let alertController = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        alertController.addAction(UIAlertAction(title: Constants.okText, style: .cancel, handler: nil))
-        present(alertController, animated: true)
     }
 }
